+++ RELEASES.md
@@ -289,3 +289,13 @@ Based on:
 - [typescript v0.9.0] .
 ### Releases
 - [NPM v0.9.0] https://www.npmjs.com/package/@lukehagar/discoursejs/v/0.9.0 - .
+
+## 2024-09-13 00:23:58
+### Changes
+Based on:
+- OpenAPI Doc latest 
+- Speakeasy CLI 1.396.6 (2.415.6) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v0.10.0] .
+### Releases
+- [NPM v0.10.0] https://www.npmjs.com/package/@lukehagar/discoursejs/v/0.10.0 - .